Do a basic memory game.
I added some icons to the memory-game public folder, so that there's
pictures to use with the memory game.

The memory game has the basic logic, but it is not complete.

- There's no official game end. You just keep selecting tiles until
there's no more tiles to match.
- There's no way to play the game more than once. No playing of a new
game, no regeneration of the board. No reset. No different difficulties.
- The codebase is a bit messy. It can be improved.

The project was done in steps, rounding things out bit-by-bit.

Changes to be committed:
      modified:   react/memory-game/src/App.js
      new file:   react/memory-game/src/steps/00-show-image/index.js
      new file:   react/memory-game/src/steps/01-show-many-images/index.js
      new file:   react/memory-game/src/steps/02-show-image-grid/index.js
      new file:   react/memory-game/src/steps/03-make-game-tile/App.css
      new file:   react/memory-game/src/steps/03-make-game-tile/index.js
      new file:   react/memory-game/src/steps/04-game-logic/App.css
      new file:   react/memory-game/src/steps/04-game-logic/index.js
      copied:     react/memory-game/src/App.js -> react/memory-game/src/steps/99-final-version/index.js