Add message display limit
Replaced how messages are stored on the backend with an array that holds
only a limited number of elements. This way, the history that a client
gets is small.

Replaced how messages are displayed on the frontend by creating a list
of `<li>` elements every time a message is received (instead of just
appending a new `<li>` element for each message).

NOTE: The `user joined` and `user left` messages aren't stored in the
backend message lists, but they're stored in the frontend message list
(for no good reason) so you're going to see some funny business between
those. How you handle that is up to you. The easiest thing to do is to
just store all of the messages in the history. Just get this done as
fast as possible.

On branch master
Your branch is up to date with 'origin/master'.

Changes to be committed:
      modified:   index.html
      modified:   index.js
      new file:   limited-array.js