Get sort visualizer to physically move elements
Moving is done through transitions of position. It hasn't been done
well, yet.

TODO:
- Make sure that the list can show up normally around other elements.
This basically means that the element which displays the list is going
to have to take over positioning elements, instead of letting the
browser do it. It will have to know how tall each of the elements is in
order to move them. The div that contains the list itself will have to
have a manually set height so that other elements can flow around the
list naturally.
- The timing for each transition is manual. Change this so that state
changes occur after all transitions end.