Add latex-base for latex commands, add new tags
Added functionality for custom tags that with numeric references. So
long as I create the tag, the numeric references will be kept. I don't
have to think about that when I create each element. The work is done
for me.

I applied this to the example tag.

Some previous tags which used latex macros changed since the way macros
work in latex-base changed as compared to how they used to work in
latex-commands.